Fixed issue where record count does not increase if new record was added
I made use of the onDialogPositiveClick callback so that a callback is
registered to the MainActivity after a record was added. Once the
callback is registered the number of records is updated and the updated
count and the new record is displayed. I also modified the existing
countRecord method so that it can return the updated record number
Signed-off-by: Alvin Dizon <alvin.dizon91@gmail.com>